Use ltdl for linking
This gives us cross-platform linking 'for-free' and doesn't
require the target have ltdl installed (--with-included_ltdl)
